Rename scribe package to zipkinscribereceiver. (#498)
